Minimizing dependency on marklogic-langchain4j module
That module is going to require Java 17. Still have a few things left to fix, but this takes care of some easy stuff by moving things into marklogic-spark-api.
