+++ marklogic-langchain4j/build.gradle
@@ -1,4 +1,6 @@
 dependencies {
+  implementation project(":marklogic-spark-api")
+
   compileOnly "com.fasterxml.jackson.core:jackson-databind:2.17.2"
   api "com.marklogic:marklogic-client-api:7.1.0"
