+++ marklogic-spark-api/build.gradle
@@ -3,5 +3,7 @@
  * it doesn't have a circular dependency with the connector subproject.
  */
 dependencies {
+  compileOnly "com.marklogic:marklogic-client-api:7.1.0"
+  compileOnly "com.fasterxml.jackson.core:jackson-databind:2.17.2"
   implementation "org.slf4j:jcl-over-slf4j:2.0.13"
 }
