+++ marklogic-spark-api/src/main/java/com/marklogic/spark/dom/DOMHelper.java
@@ -1,13 +1,13 @@
 /*
  * Copyright © 2024 MarkLogic Corporation. All Rights Reserved.
  */
-package com.marklogic.langchain4j.dom;
+package com.marklogic.spark.dom;
 
 import com.marklogic.client.document.DocumentWriteOperation;
 import com.marklogic.client.impl.HandleAccessor;
 import com.marklogic.client.io.DOMHandle;
 import com.marklogic.client.io.marker.AbstractWriteHandle;
-import com.marklogic.langchain4j.MarkLogicLangchainException;
+import com.marklogic.spark.ConnectorException;
 import org.w3c.dom.Document;
 import org.xml.sax.InputSource;
 
@@ -56,7 +56,7 @@ public class DOMHelper {
             String message = sourceUri != null ?
                 String.format("Unable to parse XML for document with URI: %s; cause: %s", sourceUri, e.getMessage()) :
                 String.format("Unable to parse XML; cause: %s", e.getMessage());
-            throw new MarkLogicLangchainException(message, e);
+            throw new ConnectorException(message, e);
         }
     }
 
@@ -74,7 +74,7 @@ public class DOMHelper {
             return xpath.compile(xpathExpression);
         } catch (XPathExpressionException e) {
             String message = massageXPathCompilationError(e.getMessage());
-            throw new MarkLogicLangchainException(String.format(
+            throw new ConnectorException(String.format(
                 "Unable to compile XPath expression for %s: %s; cause: %s",
                 purposeForErrorMessage, xpathExpression, message), e
             );
@@ -94,7 +94,7 @@ public class DOMHelper {
             try {
                 this.documentBuilder = this.documentBuilderFactory.newDocumentBuilder();
             } catch (ParserConfigurationException e) {
-                throw new MarkLogicLangchainException(String.format("Unable to create XML document; cause: %s", e.getMessage()), e);
+                throw new ConnectorException(String.format("Unable to create XML document; cause: %s", e.getMessage()), e);
             }
         }
         return this.documentBuilder;
