+++ marklogic-spark-connector/src/test/java/com/marklogic/langchain4j/embedding/EmbedderTest.java
@@ -13,9 +13,9 @@ import com.marklogic.client.impl.DocumentWriteOperationImpl;
 import com.marklogic.client.io.DocumentMetadataHandle;
 import com.marklogic.client.io.JacksonHandle;
 import com.marklogic.junit5.XmlNode;
-import com.marklogic.langchain4j.Util;
 import com.marklogic.langchain4j.splitter.*;
 import com.marklogic.spark.AbstractIntegrationTest;
+import com.marklogic.spark.Util;
 import com.marklogic.spark.writer.XmlUtil;
 import dev.langchain4j.data.document.splitter.DocumentSplitters;
 import dev.langchain4j.model.embedding.onnx.allminilml6v2.AllMiniLmL6V2EmbeddingModel;
@@ -43,7 +43,7 @@ class EmbedderTest extends AbstractIntegrationTest {
         docs.next();
 
         docs.forEachRemaining(doc -> {
-            JsonNode node = Util.getJsonFromHandle(doc.getContent());
+            JsonNode node = com.marklogic.spark.Util.getJsonFromHandle(doc.getContent());
             ArrayNode chunks = (ArrayNode) node.get("chunks");
             assertEquals(2, chunks.size());
             for (JsonNode chunk : chunks) {
