+++ marklogic-spark-langchain4j/src/main/java/com/marklogic/spark/langchain4j/NamespaceContextFactory.java
@@ -3,9 +3,9 @@
  */
 package com.marklogic.spark.langchain4j;
 
-import com.marklogic.langchain4j.Util;
-import com.marklogic.langchain4j.dom.XPathNamespaceContext;
+import com.marklogic.spark.dom.XPathNamespaceContext;
 import com.marklogic.spark.Options;
+import com.marklogic.spark.Util;
 
 import javax.xml.namespace.NamespaceContext;
 import java.util.HashMap;
@@ -18,7 +18,7 @@ public interface NamespaceContextFactory {
         prefixesToNamespaces.put("model", Util.DEFAULT_XML_NAMESPACE);
         return new XPathNamespaceContext(prefixesToNamespaces);
     }
-    
+
     static NamespaceContext makeNamespaceContext(Map<String, String> properties) {
         Map<String, String> prefixesToNamespaces = new HashMap<>();
         properties.keySet().stream()
